Clarify that cursor_pos is BYTE offset (not character)

+++ src/frame.h
@@ -128,7 +128,7 @@ extern void wrap_paragraph(Paragraph *para, PangoContext *pc, double w);
 extern size_t end_offset_of_para(struct frame *fr, int pn);
 
 extern int find_cursor(struct frame *fr, double x, double y,
-                       int *ppara, int *ppos, int *ptrail);
+                       int *ppara, size_t *ppos, int *ptrail);
 
 extern int get_para_highlight(struct frame *fr, int cursor_para,
                               double *cx, double *cy, double *cw, double *ch);
@@ -136,10 +136,10 @@ extern int get_para_highlight(struct frame *fr, int cursor_para,
 extern int get_cursor_pos(struct frame *fr, int cursor_para, int cursor_pos,
                           double *cx, double *cy, double *ch);
 
-extern void cursor_moveh(struct frame *fr, int *cpara, int *cpos, int *ctrail,
+extern void cursor_moveh(struct frame *fr, int *cpara, size_t *cpos, int *ctrail,
                          signed int dir);
 
-extern void cursor_movev(struct frame *fr, int *cpara, int *cpos, int *ctrail,
+extern void cursor_movev(struct frame *fr, int *cpara, size_t *cpos, int *ctrail,
                          signed int dir);
 
 extern void check_callback_click(struct frame *fr, int para);
